requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
python-dotenv
tiktoken

# PDF/DOCX processing
pypdf
python-docx

# LLM clients
openai
anthropic

# Embeddings (for future use)
sentence-transformers
